Pluggable Interpolation Module (#290)
